+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## Unreleased
 
+### Changed
+
+- Added a check to prevent a collection from being added as a sub-collection of itself
+
 ### Removed
 
 - **Breaking**: Removed an unused endpoint for retrieving IIIF images for a digital object and corresponding client method